Call core chart.js `destroy()` function
Calling the `destroy()` function is necessary to prevent a memory leak as otherwise the chart.js core code maintains references to every chart ever generated with this library. The `instances` variable it assigns references to, seen [here](https://github.com/chartjs/Chart.js/blob/master/src/core/core.controller.js#L105), is only ever cleared upon calling of the `destroy` function. It's not attached to any part of the DOM and therefore is not cleaned up by the destruction of the virtual DOM; it's a package-level variable that persists across all runs of the chart.js library.